# internal/alerting/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.29.0

package alerting

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type AlertInstance struct {
	InstanceID        pgtype.UUID
	MetricID          string
	Status            string
	BreachCount       int32
	RecoveryCount     int32
	NoDataCount       int32
	StateBeforeNoData pgtype.Text
	Unavailability    pgtype.Text
	UpdatedAt         pgtype.Timestamptz
}

type AppUser struct {
	ID           pgtype.UUID
	Username     string
	PasswordHash []byte
	Role         string
}

type CollectionTaskConfig struct {
	InstanceID      pgtype.UUID
	TaskID          string
	IntervalSeconds int32
	UpdatedAt       pgtype.Timestamptz
}

type Instance struct {
	ID             pgtype.UUID
	Name           string
	Host           string
	Port           int32
	DatabaseName   string
	Username       string
	Password       string
	AgentTokenHash []byte
	CreatedAt      pgtype.Timestamptz
	AgentVersion   pgtype.Text
}

type InstanceCollectState struct {
	InstanceID       pgtype.UUID
	Source           string
	LastSuccessAt    pgtype.Timestamptz
	LastReportAt     pgtype.Timestamptz
	LastErrorCode    pgtype.Text
	LastErrorMessage pgtype.Text
}

type InstanceCollectionConfig struct {
	InstanceID          pgtype.UUID
	AgentMetricsEnabled bool
	UpdatedAt           pgtype.Timestamptz
}

type MetricSample struct {
	SeriesID int64
	Ts       pgtype.Timestamptz
	Value    float64
}

type MetricSeries struct {
	SeriesID   int64
	InstanceID pgtype.UUID
	MetricID   string
	Labels     []byte
	LabelsKey  string
	FirstSeen  pgtype.Timestamptz
	LastSeen   pgtype.Timestamptz
}

type UserSession struct {
	TokenHash []byte
	UserID    pgtype.UUID
	ExpiresAt pgtype.Timestamptz
}